Rewrite graphyne-worker
