+++ packages/graphyne-worker/src/types.ts
@@ -1,0 +1,5 @@
+export type TContext = Record<string, any>;
+
+export interface HandlerConfig {
+  context?: TContext | ((request: Request) => TContext | Promise<TContext>);
+}
